# fix(p2p): make PeerSet.Remove more efficient