--- a/p2p/peer_set_test.go
+++ b/p2p/peer_set_test.go
@@ -48,13 +48,19 @@ func newMockPeer(ip net.IP) *mockPeer {
 }
 
 func TestPeerSetAddRemoveOne(t *testing.T) {
+	testPeerSetAddRemoveOne(t)
+}
+
+func testPeerSetAddRemoveOne(tb testing.TB) {
+	tb.Helper()
+
 	peerSet := NewPeerSet()
 
 	var peerList []Peer
 	for i := 0; i < 5; i++ {
 		p := newMockPeer(net.IP{127, 0, 0, byte(i)})
 		if err := peerSet.Add(p); err != nil {
-			t.Error(err)
+			tb.Error(err)
 		}
 		peerList = append(peerList, p)
 	}
@@ -63,14 +69,14 @@ func TestPeerSetAddRemoveOne(t *testing.T) {
 	// 1. Test removing from the front
 	for i, peerAtFront := range peerList {
 		removed := peerSet.Remove(peerAtFront)
-		assert.True(t, removed)
+		assert.True(tb, removed)
 		wantSize := n - i - 1
 		for j := 0; j < 2; j++ {
-			assert.False(t, false, peerSet.Has(peerAtFront.ID()), "#%d Run #%d: failed to remove peer", i, j)
-			assert.Equal(t, wantSize, peerSet.Size(), "#%d Run #%d: failed to remove peer and decrement size", i, j)
+			assert.False(tb, false, peerSet.Has(peerAtFront.ID()), "#%d Run #%d: failed to remove peer", i, j)
+			assert.Equal(tb, wantSize, peerSet.Size(), "#%d Run #%d: failed to remove peer and decrement size", i, j)
 			// Test the route of removing the now non-existent element
 			removed := peerSet.Remove(peerAtFront)
-			assert.False(t, removed)
+			assert.False(tb, removed)
 		}
 	}
 
@@ -78,7 +84,7 @@ func TestPeerSetAddRemoveOne(t *testing.T) {
 	// a) Replenish the peerSet
 	for _, peer := range peerList {
 		if err := peerSet.Add(peer); err != nil {
-			t.Error(err)
+			tb.Error(err)
 		}
 	}
 
@@ -86,13 +92,19 @@ func TestPeerSetAddRemoveOne(t *testing.T) {
 	for i := n - 1; i >= 0; i-- {
 		peerAtEnd := peerList[i]
 		removed := peerSet.Remove(peerAtEnd)
-		assert.True(t, removed)
-		assert.False(t, false, peerSet.Has(peerAtEnd.ID()), "#%d: failed to remove item at end", i)
-		assert.Equal(t, i, peerSet.Size(), "#%d: differing sizes after peerSet.Remove(atEndPeer)", i)
+		assert.True(tb, removed)
+		assert.False(tb, false, peerSet.Has(peerAtEnd.ID()), "#%d: failed to remove item at end", i)
+		assert.Equal(tb, i, peerSet.Size(), "#%d: differing sizes after peerSet.Remove(atEndPeer)", i)
 	}
 }
 
 func TestPeerSetAddRemoveMany(t *testing.T) {
+	testPeerSetAddRemoveMany(t)
+}
+
+func testPeerSetAddRemoveMany(tb testing.TB) {
+	tb.Helper()
+
 	peerSet := NewPeerSet()
 
 	peers := []Peer{}
@@ -100,22 +112,22 @@ func TestPeerSetAddRemoveMany(t *testing.T) {
 	for i := 0; i < N; i++ {
 		peer := newMockPeer(net.IP{127, 0, 0, byte(i)})
 		if err := peerSet.Add(peer); err != nil {
-			t.Errorf("failed to add new peer")
+			tb.Errorf("failed to add new peer")
 		}
 		if peerSet.Size() != i+1 {
-			t.Errorf("failed to add new peer and increment size")
+			tb.Errorf("failed to add new peer and increment size")
 		}
 		peers = append(peers, peer)
 	}
 
 	for i, peer := range peers {
 		removed := peerSet.Remove(peer)
-		assert.True(t, removed)
+		assert.True(tb, removed)
 		if peerSet.Has(peer.ID()) {
-			t.Errorf("failed to remove peer")
+			tb.Errorf("failed to remove peer")
 		}
 		if peerSet.Size() != len(peers)-i-1 {
-			t.Errorf("failed to remove peer and decrement size")
+			tb.Errorf("failed to remove peer and decrement size")
 		}
 	}
 }