@melekes 2024-02-06
Closing in favor of https://github.com/cometbft/cometbft/pull/2246